+++ graphite_api/functions.py
@@ -23,10 +23,11 @@ import math
 import re
 import random
 import six
+import time
 
 from six.moves import zip_longest, map, reduce
 
-from .render.attime import parseTimeOffset
+from .render.attime import parseTimeOffset, parseATTime
 from .render.glyph import format_units
 from .render.datalib import TimeSeries
 from .utils import to_seconds, epoch
@@ -40,6 +41,13 @@ DAY = HOUR * 24
 
 # Utility functions
 not_none = partial(filter, partial(is_not, None))
+
+
+def not_empty(values):
+    for v in values:
+        if v is not None:
+            return True
+    return False
 
 
 def safe(f):
@@ -250,16 +258,60 @@ def averageSeriesWithWildcards(requestContext, seriesList, *positions):
     return result
 
 
+def multiplySeriesWithWildcards(requestContext, seriesList, *position):
+    """
+    Call multiplySeries after inserting wildcards at the given position(s).
+
+    Example::
+
+        &target=multiplySeriesWithWildcards(
+            web.host-[0-7].{avg-response,total-request}.value, 2)
+
+    This would be the equivalent of::
+
+        &target=multiplySeries(web.host-0.{avg-response,total-request}.value)
+        &target=multiplySeries(web.host-1.{avg-response,total-request}.value)
+        ...
+    """
+    positions = [position] if isinstance(position, int) else position
+
+    newSeries = {}
+    newNames = []
+
+    for series in seriesList:
+        new_name = ".".join(map(lambda x: x[1],
+                                filter(lambda i: i[0] not in positions,
+                                       enumerate(series.name.split('.')))))
+
+        if new_name in newSeries:
+            [newSeries[new_name]] = multiplySeries(requestContext,
+                                                   (newSeries[new_name],
+                                                    series))
+        else:
+            newSeries[new_name] = series
+            newNames.append(new_name)
+        newSeries[new_name].name = new_name
+    return [newSeries[name] for name in newNames]
+
+
 def diffSeries(requestContext, *seriesLists):
     """
-    Can take two or more metrics.
-    Subtracts parameters 2 through n from parameter 1.
+    Subtracts series 2 through n from series 1.
 
     Example::
 
         &target=diffSeries(service.connections.total,
                            service.connections.failed)
 
+    To diff a series and a constant, one should use offset instead of
+    (or in addition to) diffSeries.
+
+    Example::
+
+        &target=offset(service.connections.total, -5)
+
+        &target=offset(diffSeries(service.connections.total,
+                                  service.connections.failed), -4)
     """
     if not seriesLists or seriesLists == ([],):
         return []
@@ -1235,10 +1287,11 @@ def cactiStyle(requestContext, seriesList, system=None):
         &target=cactiStyle(ganglia.*.net.bytes_out,"si")
 
     """
-    if system:
-        fmt = lambda x: "%.2f%s" % format_units(x, system=system)
-    else:
-        fmt = lambda x: "%.2f" % x
+    def fmt(x):
+        if system:
+            return "%.2f%s" % format_units(x, system=system)
+        else:
+            return "%.2f" % x
     nameLen = max([0] + [len(series.name) for series in seriesList])
     lastLen = max([0] + [len(fmt(int(safeLast(series) or 3)))
                          for series in seriesList]) + 3
@@ -1978,8 +2031,8 @@ def stdev(requestContext, seriesList, points, windowTolerance=0.1):
                 float(validPoints) / points >= windowTolerance
             ):
 
-                deviation = math.sqrt(validPoints * currentSumOfSquares
-                                      - currentSum**2) / validPoints
+                deviation = math.sqrt(validPoints * currentSumOfSquares -
+                                      currentSum**2) / validPoints
                 stddevSeries.append(deviation)
             else:
                 stddevSeries.append(None)
@@ -2055,8 +2108,8 @@ def _trimBootstrap(bootstrap, original):
 
 def holtWintersIntercept(alpha, actual, last_season, last_intercept,
                          last_slope):
-    return (alpha * (actual - last_season)
-            + (1 - alpha) * (last_intercept + last_slope))
+    return (alpha * (actual - last_season) +
+            (1 - alpha) * (last_intercept + last_slope))
 
 
 def holtWintersSlope(beta, intercept, last_intercept, last_slope):
@@ -2070,8 +2123,8 @@ def holtWintersSeasonal(gamma, actual, intercept, last_season):
 def holtWintersDeviation(gamma, actual, prediction, last_seasonal_dev):
     if prediction is None:
         prediction = 0
-    return (gamma * math.fabs(actual - prediction)
-            + (1 - gamma) * last_seasonal_dev)
+    return (gamma * math.fabs(actual - prediction) +
+            (1 - gamma) * last_seasonal_dev)
 
 
 def holtWintersAnalysis(series):
@@ -2332,8 +2385,8 @@ def timeStack(requestContext, seriesList, timeShiftUnit, timeShiftStart,
     with time shifts starting time shifts from the start multiplier through
     the end multiplier.
 
-    Useful for looking at history, or feeding into seriesAverage or
-    seriesStdDev.
+    Useful for looking at history, or feeding into averageSeries or
+    stddevSeries.
 
     Example::
 
@@ -2422,6 +2475,38 @@ def timeShift(requestContext, seriesList, timeShift, resetEnd=True):
         shiftedSeries.start = series.start
         results.append(shiftedSeries)
 
+    return results
+
+
+def timeSlice(requestContext, seriesList, startSliceAt, endSliceAt='now'):
+    """
+    Takes one metric or a wildcard metric, followed by a quoted
+    string with the time to start the line and another quoted string
+    with the time to end the line. The start and end times are
+    inclusive. See ``from / until`` in the render api for examples of
+    time formats.
+
+    Useful for filtering out a part of a series of data from a wider
+    range of data.
+
+    Example::
+
+        &target=timeSlice(network.core.port1,"00:00 20140101","11:59 20140630")
+        &target=timeSlice(network.core.port1,"12:00 20140630","now")
+    """
+    results = []
+    start = time.mktime(parseATTime(startSliceAt).timetuple())
+    end = time.mktime(parseATTime(endSliceAt).timetuple())
+
+    for slicedSeries in seriesList:
+        slicedSeries.name = 'timeSlice(%s, %s, %s)' % (slicedSeries.name,
+                                                       int(start), int(end))
+        curr = time.mktime(requestContext["startTime"].timetuple())
+        for i, v in enumerate(slicedSeries):
+            if v is None or curr < start or curr > end:
+                slicedSeries[i] = None
+            curr += slicedSeries.step
+        results.append(slicedSeries)
     return results
 
 
@@ -2559,7 +2644,7 @@ def isNonNull(requestContext, seriesList):
     return seriesList
 
 
-def identity(requestContext, name):
+def identity(requestContext, name, step=60):
     """
     Identity function:
     Returns datapoints where the value equals the timestamp of the datapoint.
@@ -2572,8 +2657,10 @@ def identity(requestContext, name):
 
     This would create a series named "The.time.series" that contains points
     where x(t) == t.
+
+    Accepts optional second argument as 'step' parameter (default step is
+    60 sec)
     """
-    step = 60
     start = int(epoch(requestContext["startTime"]))
     end = int(epoch(requestContext["endTime"]))
     values = range(start, end, step)
@@ -2795,14 +2882,17 @@ def smartSummarize(requestContext, seriesList, intervalString, func='sum'):
 
     # Adjust the start time to fit an entire day for intervals >= 1 day
     requestContext = requestContext.copy()
+    tzinfo = requestContext['tzinfo']
     s = requestContext['startTime']
     if interval >= DAY:
-        requestContext['startTime'] = datetime(s.year, s.month, s.day)
+        requestContext['startTime'] = datetime(s.year, s.month, s.day,
+                                               tzinfo=tzinfo)
     elif interval >= HOUR:
-        requestContext['startTime'] = datetime(s.year, s.month, s.day, s.hour)
+        requestContext['startTime'] = datetime(s.year, s.month, s.day, s.hour,
+                                               tzinfo=tzinfo)
     elif interval >= MINUTE:
         requestContext['startTime'] = datetime(s.year, s.month, s.day, s.hour,
-                                               s.minute)
+                                               s.minute, tzinfo=tzinfo)
 
     for i, series in enumerate(seriesList):
         # XXX: breaks with summarize(metric.{a,b})
@@ -2822,6 +2912,10 @@ def smartSummarize(requestContext, seriesList, intervalString, func='sum'):
 
         # Populate buckets
         for timestamp, value in datapoints:
+            # ISSUE: Sometimes there is a missing timestamp in datapoints when
+            #        running a smartSummary
+            if not timestamp:
+                continue
             bucketInterval = int((timestamp - series.start) / interval)
 
             if bucketInterval not in buckets:
@@ -2909,11 +3003,13 @@ def summarize(requestContext, seriesList, intervalString, func='sum',
     for series in seriesList:
         buckets = {}
 
-        timestamps = range(int(series.start), int(series.end),
+        timestamps = range(int(series.start), int(series.end) + 1,
                            int(series.step))
         datapoints = zip_longest(timestamps, series)
 
         for timestamp, value in datapoints:
+            if timestamp is None:
+                continue
             if alignToFrom:
                 bucketInterval = int((timestamp - series.start) / interval)
             else:
@@ -3061,7 +3157,7 @@ def hitcount(requestContext, seriesList, intervalString,
     return results
 
 
-def sinFunction(requestContext, name, amplitude=1):
+def sinFunction(requestContext, name, amplitude=1, step=60):
     """
     Short Alias: sin()
 
@@ -3073,8 +3169,9 @@ def sinFunction(requestContext, name, amplitude=1):
         &target=sin("The.time.series", 2)
 
     This would create a series named "The.time.series" that contains sin(x)*2.
+
+    A third argument can be provided as a step parameter (default is 60 secs).
     """
-    step = 60
     delta = timedelta(seconds=step)
     when = requestContext["startTime"]
     values = []
@@ -3091,7 +3188,21 @@ def sinFunction(requestContext, name, amplitude=1):
     return [series]
 
 
-def randomWalkFunction(requestContext, name):
+def removeEmptySeries(requestContext, seriesList):
+    """
+    Takes one metric or a wildcard seriesList. Out of all metrics
+    passed, draws only the metrics with not empty data.
+
+    Example::
+
+        &target=removeEmptySeries(server*.instance*.threads.busy)
+
+    Draws only live servers with not empty data.
+    """
+    return [series for series in seriesList if not_empty(series)]
+
+
+def randomWalkFunction(requestContext, name, step=60):
     """
     Short Alias: randomWalk()
 
@@ -3104,8 +3215,10 @@ def randomWalkFunction(requestContext, name):
 
     This would create a series named "The.time.series" that contains points
     where x(t) == x(t-1)+random()-0.5, and x(0) == 0.
+
+    Accepts an optional second argument as step parameter (default step is
+    60 sec).
     """
-    step = 60
     delta = timedelta(seconds=step)
     when = requestContext["startTime"]
     values = []
@@ -3149,6 +3262,7 @@ SeriesFunctions = {
     'avg': averageSeries,
     'sumSeriesWithWildcards': sumSeriesWithWildcards,
     'averageSeriesWithWildcards': averageSeriesWithWildcards,
+    'multiplySeriesWithWildcards': multiplySeriesWithWildcards,
     'minSeries': minSeries,
     'maxSeries': maxSeries,
     'rangeOfSeries': rangeOfSeries,
@@ -3170,6 +3284,7 @@ SeriesFunctions = {
     'log': logarithm,
     'timeStack': timeStack,
     'timeShift': timeShift,
+    'timeSlice': timeSlice,
     'summarize': summarize,
     'smartSummarize': smartSummarize,
     'hitcount': hitcount,
@@ -3213,6 +3328,7 @@ SeriesFunctions = {
     'useSeriesAbove': useSeriesAbove,
     'exclude': exclude,
     'grep': grep,
+    'removeEmptySeries': removeEmptySeries,
 
     # Data Filter functions
     'removeAbovePercentile': removeAbovePercentile,
@@ -3240,7 +3356,9 @@ SeriesFunctions = {
     'substr': substr,
     'group': group,
     'map': mapSeries,
+    'mapSeries': mapSeries,
     'reduce': reduceSeries,
+    'reduceSeries': reduceSeries,
     'groupByNode': groupByNode,
     'constantLine': constantLine,
     'stacked': stacked,